Ensure that we scan @Provides-like methods in a consistent ordering.  Previously we used the results of Class.getDeclaredMethods as-is, which can be non-deterministic from run to run on some JVMs.  This will allow some tests to have more stable behavior.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=235046189
